refactor: change menu page folder structure
